+++ cpskin/core/events.py
@@ -1,8 +1,11 @@
-from plone import api
-
-from Products.CMFCore.utils import getToolByName
-
+# -*- coding: utf-8 -*-
 from cpskin.core.interfaces import ICPSkinCoreWithMembersLayer
+from cpskin.core.utils import get_address_from_obj
+from cpskin.core.utils import get_lat_lng_from_address
+from cpskin.core.utils import has_lat_lng
+from cpskin.core.utils import set_coord
+from plone import api
+from Products.CMFCore.utils import getToolByName
 
 
 def user_logged_in(event):
@@ -60,3 +63,14 @@ def user_initial_logged_in(event):
     home.setConstrainTypesMode(1)
     home.setLocallyAllowedTypes(local_roles)
     home.setImmediatelyAddableTypes(local_roles)
+
+
+def set_lat_lng(obj, event):
+    """Set lat and lng when a organization is created"""
+    if has_lat_lng(obj):
+        return
+    address = get_address_from_obj(obj)
+    if not address:
+        return
+    request = obj.REQUEST
+    set_coord(obj, request)